Restructure S3 file path
The s3 file path is currently constructed so that the timestamp
is added to the end of the file name. This is causing issues where
the file's organogram visualisation isn't showing, and is expecting
a file name that ends in `-senior.csv` or `-junior.csv`.

The timestamp has now been moved to the beginning of the filename,
and if it's not provided by the user in the `resource` it will use
today's current timestamp.

Moving the timestamp to the beginning of the file name should fix
this.